third_party/WebKit/LayoutTests/typedcssom/inlinestyle/numbers-expected.txt

This is a testharness.js-based test.
FAIL Single valued CSSUnitValues can be retrieved from Inline StyleMap assert_equals: result from undefined expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can retrieve list of CSSUnitValues from list-valued property assert_equals: expected 3 but got 1
Harness: the test ran to completion.
